+++ Changelog.md
@@ -6,6 +6,9 @@ This file holds the changelog for Pepperminty Wiki. This is the master list of t
 ## Added
  - Watchlists! A new addition has been added to the more menu to add the current page to your personal watchlist
      - An email will be sent to all users watching a page when an edit is saved to it (uses the PHP `mail()` function internally, via the [`email_user()`](https://starbeamrainbowlabs.com/labs/peppermint/docs/ModuleApi/#method_email_user) internal Pepperminty Wiki utility function)
+ - Email address verification
+     - Enabled by default. In order to receive emails users now need to verify their email address
+     - This is done via a verification email that's sent when you change your email address (even if your email address is the same when you change your preferences and you haven't yet verified it)
 
 ## Fixed
  - Fixed weighted word support on search query analysis debug page